+++ pynenc/runner/__init__.py
@@ -1,4 +1,5 @@
 from pynenc.runner.base_runner import BaseRunner, DummyRunner
+from pynenc.runner.multi_thread_runner import MultiThreadRunner
 from pynenc.runner.process_runner import ProcessRunner
 from pynenc.runner.thread_runner import ThreadRunner
 
@@ -7,4 +8,5 @@ __all__ = [
     "DummyRunner",
     "ThreadRunner",
     "ProcessRunner",
+    "MultiThreadRunner",
 ]